Fixed example project, where UITableView cornerRadius and separatorHeight assigned incorrectly.
When UITableView first time loaded, cornerRadius and separatorHeight is
drawn correct. But, when you scroll down and force UITableViewCell
redraw, cornerRadius and separatorHeight is set to initial values from
initialize method.  Maybe it's because of defective implementation in
FUICellBackgroundView or UITableViewCell+FlatUI ?! At least this commit
fixes the example project demonstration correctly.